fix bugs

+++ modules/v1/users/users.controller.go
@@ -99,7 +99,9 @@ func (c *user_ctrl) AddUser(w http.ResponseWriter, r *http.Request) {
 	_, err = govalidator.ValidateStruct(data)
 
 	if err != nil {
-		_ = os.Remove(imageName)
+		if imageName != "public/default_image.jpg" {
+			_ = os.Remove(imageName)
+		}
 		libs.GetResponse(err.Error(), 400, true).Send(w)
 		return
 	}
@@ -128,7 +130,9 @@ func (c *user_ctrl) UpdateUser(w http.ResponseWriter, r *http.Request) {
 	_, err = govalidator.ValidateStruct(data)
 
 	if err != nil {
-		_ = os.Remove(imageName)
+		if imageName != "public/default_image.jpg" {
+			_ = os.Remove(imageName)
+		}
 		libs.GetResponse(err.Error(), 400, true).Send(w)
 		return
 	}
